Add Account Aggregate, Refactor Event
