+++ account-api/account-domain/src/main/java/kr/flab/movieon/account/domain/event/RegisteredAccountConfirmEvent.java
@@ -1,6 +1,7 @@
-package kr.flab.movieon.account.domain;
+package kr.flab.movieon.account.domain.event;
 
 import java.util.Date;
+import kr.flab.movieon.account.domain.Account;
 import kr.flab.movieon.common.DomainEvent;
 import lombok.Getter;
 
@@ -9,14 +10,13 @@ public final class RegisteredAccountConfirmEvent implements DomainEvent {
 
     private final Long accountId;
     private final String email;
-    private final String emailCheckToken;
+    private final String emailValidationToken;
     private final Date occurredOn;
 
-    public RegisteredAccountConfirmEvent(Long accountId, String email,
-        String emailCheckToken) {
-        this.accountId = accountId;
-        this.email = email;
-        this.emailCheckToken = emailCheckToken;
+    public RegisteredAccountConfirmEvent(Account account) {
+        this.accountId = account.getId();
+        this.email = account.getEmail();
+        this.emailValidationToken = account.getEmailValidationToken();
         this.occurredOn = new Date();
     }
