+++ account-api/account-domain/src/test/java/kr/flab/movieon/account/domain/RegisterAccountConfirmProcessorTest.java
@@ -22,7 +22,7 @@ final class RegisterAccountConfirmProcessorTest {
         var accountRepository = new FakeAccountRepository();
         var processor = new RegisterAccountConfirmProcessor(accountRepository);
 
-        var account = Account.of("jiwon", "jiwon@gmail.com", "password!");
+        var account = Account.create("jiwon", "jiwon@gmail.com", "password!");
         accountRepository.save(account);
 
         return Arrays.asList(
